+++ src/components/PostList/DefaultPostList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './DefaultPostList';
