+++ src/components/PostList/TileViewPostList/TileViewItem.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { TilePostProps } from 'types';
 import { timeForToday } from 'lib/common';
-import styles from './ListViewItem.module.scss';
+import styles from './TileViewPostList.module.scss';
 
 interface props {
   value: TilePostProps;
