+++ src/components/PostList/TileViewPostList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TileViewPostList';
